HumanReadableDuration now implements TemporalAmount
TemporalAmount is the Java 8 interface for an 'amount of time'. A lot of
people internally use HumanReadableDuration pervasively; it effectively
replaces the use of the Java 8 Duration class.

If we'd wanted otherwise, we shouldn't have added any accessors other
than 'toJavaDuration'.

This PR makes it implement TemporalAmount, which means you can be like
Instant.now().plus(hrd);