Apply the Jetbrains Pattern annotation to the QosReason ctor
This provides immediate IDE feedback when values are passed that
don't satisfy our regex. This pattern is helpful for constant-string
inputs, but requires additional IDE config if we begin to use it
elsewhere that constants aren't required (e.g. ErrorType names)
unless we disable the editor validation that all inputs can be
traced to a constant value.