faster parsing for primary name component of user-agent strings
Adds a method `UserAgents#tryParsePrimaryName`, which attempts to parse
and return _only_ the "name" component of the primary user-agent, and
ignores everything else. It tries to be faster by explicitly not doing
any regular expression matching and only examining a subset of the
user-agent string (up to the very first "/" character).